# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: BACnet/IP REST Proxy
  description: >
    Translates HTTP REST calls into BACnet/IP requests against configured
    BACnet devices. Access is controlled by bearer-token authentication and a
    rule-based authorization engine (see the project README).
  version: "1.0.0"
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /devices/{deviceId}/objects/{objectType}/{instance}/{property}:
    get:
      tags: [properties]
      summary: Read one property
      operationId: getProperty
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - $ref: "#/components/parameters/ObjectType"
        - $ref: "#/components/parameters/Instance"
        - $ref: "#/components/parameters/Property"
        - name: index
          in: query
          description: >
            The BACnet property array index: a 1-based index into an
            array-valued property (e.g. one slot of Priority_Array, or one
            entry of Object_List), read or written on its own instead of the
            whole array. Omit to address the whole property.
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: The property's value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyValue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BACnetError"
        "504":
          $ref: "#/components/responses/BACnetTimeout"
    put:
      tags: [properties]
      summary: Write one property
      description: >
        Issues a BACnet WriteProperty request. Requires readwrite
        permission. A null value relinquishes the given priority (default
        16) on a commandable property, per BACnet's command prioritization
        scheme (see WriteRequest.priority below).
      operationId: writeProperty
      parameters:
        - $ref: "#/components/parameters/DeviceId"
        - $ref: "#/components/parameters/ObjectType"
        - $ref: "#/components/parameters/Instance"
        - $ref: "#/components/parameters/Property"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteRequest"
      responses:
        "204":
          description: Write succeeded.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Value rejected by the device (e.g. out of range).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          $ref: "#/components/responses/BACnetError"
        "504":
          $ref: "#/components/responses/BACnetTimeout"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Either an opaque token from authentication.tokens, or a JWT signed
        with the configured shared secret.
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      description: A configured device alias, or a literal hostname/IP[:port].
      schema:
        type: string
    ObjectType:
      name: objectType
      in: path
      required: true
      example: analog-input
      description: |
        The BACnet object type, identifying what kind of object this is
        (ASHRAE 135 clause 12). Together with `instance`, this forms the
        object's BACnetObjectIdentifier.

        - `analog-input` — a physical or calculated analog (real-number)
          input point, e.g. a temperature or pressure sensor.
        - `analog-output` — a physical analog output point that can be
          commanded, e.g. a valve or damper position.
        - `analog-value` — a generic analog value not tied to a specific
          physical I/O point, e.g. a setpoint or calculated value.
        - `binary-input` — a physical or calculated two-state
          (active/inactive) input point, e.g. a status switch.
        - `binary-output` — a physical two-state output point that can be
          commanded, e.g. a relay, fan, or start/stop point.
        - `binary-value` — a generic two-state value not tied to a specific
          physical I/O point.
        - `device` — the device object itself: every BACnet device exposes
          exactly one, representing the device as a whole.
        - `multi-state-input` — an input point whose value is one of a
          discrete, ordered set of states (an unsigned state index), e.g. a
          mode indicator.
        - `multi-state-output` — a commandable output point whose value is
          one of a discrete set of states.
        - `multi-state-value` — a generic multi-state value not tied to a
          specific physical I/O point.
      schema:
        type: string
        enum:
          - analog-input
          - analog-output
          - analog-value
          - binary-input
          - binary-output
          - binary-value
          - device
          - multi-state-input
          - multi-state-output
          - multi-state-value
    Instance:
      name: instance
      in: path
      required: true
      description: >
        The object's instance number (0-4194302), unique among objects of
        this `objectType` on the device. Forms the low 22 bits of the
        object's BACnetObjectIdentifier.
      schema:
        type: integer
        minimum: 0
    Property:
      name: property
      in: path
      required: true
      example: present-value
      description: |
        The BACnet property identifier to read or write (ASHRAE 135
        clause 12 defines which properties each object type supports).
        Recognized values:

        - `object-identifier` — the object's own BACnetObjectIdentifier
          (type + instance).
        - `object-name` — a user-visible name for the object, unique among
          all objects on the device.
        - `object-type` — the object's BACnet object type enumeration.
        - `present-value` — the current value of the object's primary
          variable: a real number for Analog objects, an active/inactive
          state for Binary objects, or a 1-based state index for
          Multi-state objects. Writable, via the priority array, on
          commandable (Output/Value) objects.
        - `description` — a free-text, human-readable description of the
          object, e.g. its physical location or purpose.
        - `status-flags` — a 4-bit BACnet BIT STRING, `[in-alarm, fault,
          overridden, out-of-service]`, summarizing the object's
          operational health.
        - `event-state` — the object's current event/alarm state (e.g.
          `normal`, `fault`, `off-normal`).
        - `out-of-service` — whether Present_Value has been decoupled from
          the physical input/output it represents, e.g. for testing.
        - `units` — the engineering units Present_Value is expressed in
          (BACnet's Engineering_Units enumeration, e.g.
          `degrees-celsius`, `percent`); Analog objects only.
        - `priority-array` — the 16-slot command priority array backing a
          commandable Present_Value; each slot holds the value written at
          that priority (1 = highest, e.g. manual-life-safety) or is
          `null` if that priority is relinquished. The highest-priority
          (lowest-numbered) non-null slot determines the effective
          Present_Value.
        - `relinquish-default` — the Present_Value used when every
          priority-array slot is relinquished (`null`).
        - `object-list` — the list of BACnetObjectIdentifiers for every
          object on the device; meaningful only on the device object.
        - `vendor-name` — the device manufacturer's name; device object
          only.
        - `model-name` — the manufacturer's model name/number; device
          object only.
        - `firmware-revision` — the device's firmware revision string;
          device object only.
        - `application-software-version` — the device's application
          software version string; device object only.
        - `number-of-states` — the number of discrete states a
          Multi-state object supports (states are numbered 1..N).
      schema:
        type: string
  schemas:
...
    Access:
      type: string
      enum: [readonly, readwrite]
      description: >
        The caller's effective right on this object (variable), from the
        authorization rules: `readonly` if it may read but not write it,
        `readwrite` if it may also write it. Objects the caller may not
        read at all are never returned. This reflects the proxy's
        authorization only; the device may still reject a write, e.g. to
        a property that isn't writable.
...
    PropertyValue:
      type: object
      properties:
        value:
          description: >
            The property's value, translated from its BACnet primitive
            encoding to JSON: Real/Unsigned/Signed Integer as a number,
            Boolean as a bool, Enumerated as its numeric value,
            Character String as a string, Date/Time as a string, Bit
            String as an array of booleans, Object Identifier as
            `{type, instance}`, and an array-valued property (e.g.
            Object_List, Priority_Array) as a JSON array of the above.
            A device object's Object_List only contains the objects the
            caller may read, and `?index=` is applied to that filtered
            list (index 0 is its length).
        access:
          $ref: "#/components/schemas/Access"
    WriteRequest:
      type: object
      required: [value]
      properties:
        value:
          nullable: true
          description: >
            The value to write, in the same JSON shape PropertyValue.value
            uses for reads. `null` relinquishes the given priority (or
            priority 16 if omitted) on a commandable property, per
            BACnet's WriteProperty semantics — it does not write a literal
            null, it clears that slot of the Priority_Array.
        priority:
          type: integer
          minimum: 1
          maximum: 16
          description: >
            The BACnet command priority (1-16) to write at, for
            commandable properties (Present_Value on Output/Value
            objects). BACnet defines 16 priority levels, 1 being highest
            (e.g. manual-life-safety) and 16 lowest (manual-operator /
            default); the value in the lowest-numbered occupied slot of
            the Priority_Array wins. Defaults to 16 if omitted. Ignored —
            and rejected with write-access-denied — for non-commandable
            properties.
    Error:
      type: object
      properties:
        error:
          type: string
          description: A human-readable description of what went wrong.
        bacnetErrorClass:
          type: string
          description: >
            The numeric BACnet Error_Class (ASHRAE 135 clause 21) if the
            device itself rejected the request, identifying the general
            category of failure: `0` device, `1` object, `2` property,
            `3` resources, `4` security, `5` services, `7` communication.
        bacnetErrorCode:
          type: string
          description: >
            The numeric BACnet Error_Code (ASHRAE 135 clause 21) giving
            the specific reason, e.g. `30` timeout, `31` unknown-object,
            `32` unknown-property, `37` value-out-of-range, `39`
            write-access-denied.
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: No credentials were supplied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Credentials were supplied but access was denied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Unknown object or property.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BACnetError:
      description: The device returned a BACnet error, or a protocol-level Reject/Abort.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BACnetTimeout:
      description: The device did not respond in time.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"